requirements.txt
certifi==2025.1.31
charset-normalizer>=2.0.0,<3.0.0
idna==3.10
oauthlib==3.2.2
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.1.0
requests==2.28.0
requests-oauthlib==2.0.0
tweepy==4.15.0
update-checker==0.18.0
urllib3>=1.21.1,<1.27
websocket-client==1.8.0
google-api-python-client
pandas
kafka-python
